TODO B: Review the HTML structure
Review the HTML structure and make sure that it is valid and contains required elements. Edit and re-organize the HTML as needed.
Clean up the following code so that it's easier to see the different routes and handlers for the API, and simpler to add new ones.
Ensure that the page is responsive for tablet and mobile.
When typing in the "title" field, we want to auto-complete based on article titles that already exist.

Todo B: #3
Change-Id: